# kurae-api/internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Kurae API
  description: REST API for Kurae — limited drops, inventory, checkout, and seller dashboard.
  version: 0.1.0
  contact:
    name: Kurae
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      tags: [auth]
      summary: Seller login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Error"
...
  /auth/profile:
    patch:
      tags: [auth]
      summary: Update seller brand name
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/password:
    patch:
      tags: [auth]
      summary: Change seller password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password changed
        "401":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        slug:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        ok:
          type: boolean
        session:
          $ref: "#/components/schemas/SellerSession"
        token:
          type: string
    SellerSession:
      type: object
      properties:
        email:
          type: string
        sellerSlug:
          type: string
        sellerName:
          type: string
...